store item info to mongodb, pic store folder associated with url hash

+++ pinterest/pipelines.py
@@ -11,6 +11,32 @@ import scrapy
 import hashlib
 import shutil
 from scrapy.contrib.pipeline.images import ImagesPipeline
+import pymongo
+from scrapy.conf import settings
+from pymongo import MongoClient
+from scrapy import log
+
+
+class MongoPipeline(object):
+    def __init__(self):
+        connection = MongoClient(
+            settings['MONGODB_SERVER'],
+            settings['MONGODB_PORT']
+        )
+        db = connection[settings['MONGODB_DB']]
+        self.collection = db[settings['MONGODB_COLLECTION']]
+
+    def process_item(self, item, spider):
+        valid = True
+        for data in item:
+            if not data:
+                valid = False
+                raise DropItem('Missing{0}!'.format(data))
+        if valid:
+            self.collection.insert(dict(item))
+            log.msg('question added to mongodb database!',
+                    level=log.DEBUG, spider=spider)
+        return item
 
 
 class MyImagesPipeline(ImagesPipeline):
@@ -23,16 +49,22 @@ class MyImagesPipeline(ImagesPipeline):
 
     def item_completed(self, results, item, info):
         image_ori_paths = [x['path'] for ok, x in results if ok]
-        imgFolder = "D:/Artwork/full/" + item["top_tag"] + '/'  # 下载图片的保存路径
+        image_filename = image_ori_paths[0].split('/')[-1]
+        imgFolder = settings['FULL_PIC_STORE'] + image_filename[0:2] + '/' + image_filename[2:4] + '/' + image_filename[4:6] + '/'  # 下载图片的保存路径
         if not os.path.isdir(imgFolder):
             os.makedirs(imgFolder)
-        origin = sys.path[0] + '/Artwork/' + image_ori_paths[0]
-        new_classified_path = imgFolder + image_ori_paths[0].split('/')[-1]
         if not image_ori_paths:
             raise DropItem("Item contains no images")
-        shutil.move(origin, new_classified_path)
-        item['image_paths'] = image_ori_paths
+        else:
+            origin = sys.path[0] + '/Artwork/' + image_ori_paths[0]
+            new_classified_path = imgFolder + image_filename
+            if os.path.isfile(origin):
+                shutil.move(origin, new_classified_path)
+                item['image_paths'] = new_classified_path
         return item
+
+
+
 
 # Deprecated ImagePipeline
 class ImageDownloadPipeline(ImagesPipeline):
